+++ filmsyl-frontend/.python-version
@@ -1,0 +1,1 @@
+films-you-like
