+++ filmsyl-frontend/app.py
@@ -3,6 +3,11 @@ import requests
 from datetime import datetime
 from data_source import call_api
 
+#base_url = f'https://taxifare.lewagon.ai/predict'
+base_url = 'http://127.0.0.1:8000'
+params = {}
+
+clusters = call_api(f'{base_url}/cluster', params)
 
 st.set_page_config(
             page_title="Quick reference", # => Quick reference - Streamlit
@@ -15,11 +20,13 @@ st.set_page_config(
 introduction
 '''
 
-
 categorisation, netflix, cinemas = st.tabs(["Film categorisation", "Netflix upload", "Movie recommendation"])
 with categorisation:
-   st.header("A cat")
-   st.image("https://static.streamlit.io/examples/cat.jpg", width=200)
+   for i, cluster in enumerate(clusters):
+       st.header(f"cluster {i}")
+       for movie in cluster:
+           st.text(movie)
+
 
 with netflix:
    st.header("A dog")
@@ -30,17 +37,6 @@ with cinemas:
    st.image("https://static.streamlit.io/examples/owl.jpg", width=200)
 
 
-base_url = f'https://taxifare.lewagon.ai/predict'
+
 
 #?pickup_latitude=pi&pickup_longitude=-82.5359751617647&dropoff_latitude=40.720201&dropoff_longitude=-74.032574&passenger_count=1&pickup_datetime=2024-03-01%2016:08:41
-
-params = {
-    'pickup_latitude': 'pickup_lat',
-    'pickup_longitude': 'pickup_lon',
-    'dropoff_latitude':' dropoff_lat',
-    'dropoff_longitude': 'dropoff_lon',
-    'passenger_count': 'passengers',
-    #'pickup_datetime': f'{'pickup_date'}%{'pickup_time'}'
-}
-call_api(base_url, params)
-requests.get(base_url,  params=params)
