--- filmsyl-website/requirements.txt
@@ -1,3 +1,0 @@
-# packaging
-streamlit
-pylint
