--- requirements.txt
@@ -1,36 +1,0 @@
-# packaging
-pip>=24
-#setuptools>=26
-#twine
-#wheel>=0.29
-
-#python
-ipdb
-jupyterlab
-
-# data science
-numpy
-pandas
-scikit-learn
-matplotlib
-seaborn
-
-# tests/linter
-#black
-coverage
-#flake8
-pytest
-#yapf
-pylint
-
-# API
-gcsfs
-google-cloud-storage
-#mlflow
-#s3fs
-
-# utilities
-#six>=1.14
-#joblib
-#memoized-property
-#termcolor
